# docs/openapi.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "enum": [
                  "validation_error",
                  "bad_request",
                  "unauthorized",
                  "not_found",
                  "method_not_allowed",
                  "backend_unavailable",
                  "internal_error"
                ],
                "type": "string"
              },
              "details": {
                "items": {
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                },
                "type": "array"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "SuccessEnvelope": {
        "properties": {
          "data": {
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "meta": {
            "properties": {
              "has_more": {
                "type": "boolean"
              },
              "next_cursor": {
                "type": "string"
              }
            },
            "required": [
              "has_more"
            ],
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Bearer \u003ckey\u003e from NFC_API_KEYS",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "NetFlow Collector API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/exporters": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Success envelope"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed query string or undecodable cursor"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or unknown API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Well-formed request, semantically invalid parameter"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The storage backend did not answer"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Every known exporter with first/last seen and label."
      }
    },
    "/v1/flows": {
      "get": {
        "parameters": [
          {
            "description": "Start of the half-open range [start, end) on received_at, RFC 3339 UTC.",
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "End of the range, RFC 3339 UTC. Must be strictly after start.",
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Only records from this exporter id.",
            "in": "query",
            "name": "exporter_id",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Exact-match source address, IPv4 or IPv6.",
            "in": "query",
            "name": "src_addr",
            "required": false,
            "schema": {
              "format": "ip",
              "type": "string"
            }
          },
          {
            "description": "Exact-match destination address, IPv4 or IPv6.",
            "in": "query",
            "name": "dst_addr",
            "required": false,
            "schema": {
              "format": "ip",
              "type": "string"
            }
          },
          {
            "description": "IANA protocol number: 6 TCP, 17 UDP, 1 ICMP.",
            "in": "query",
            "name": "protocol",
            "required": false,
            "schema": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Records per page. Defaults to 100 and is clamped to 1000; a value below 1 means the default.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous response's meta.next_cursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            },
            "description": "Success envelope"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed query string or undecodable cursor"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or unknown API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Well-formed request, semantically invalid parameter"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The storage backend did not answer"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Cursor-paginated flow records over a bounded time range, with optional filters."
      }
    }
  }
}